added id validation for UserController

+++ src/main/java/com/staskost/jdbc/controller/UserController.java
@@ -3,6 +3,7 @@ package com.staskost.jdbc.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.staskost.jdbc.dao.UserDao;
 import com.staskost.jdbc.model.User;
@@ -30,7 +32,11 @@ public class UserController {
 
 	@GetMapping("/{id}")
 	public User getUserById(@PathVariable int id) {
-		return userDao.getUserById(id);
+		User user = userDao.getUserById(id);
+		if (user == null) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User Not Found");
+		}
+		return user;
 	}
 
 	@PostMapping
@@ -45,6 +51,10 @@ public class UserController {
 
 	@DeleteMapping("/{id}")
 	public @ResponseBody void deleteUser(@PathVariable int id) {
+		User user = userDao.getUserById(id);
+		if (user == null) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User Not Found");
+		}
 		userDao.deleteUser(id);
 	}
